reorganize filter code into smaller files
